1. Added game logic 2. Put all files into default package for now
1. I have three classes for the logic: GameLogicBoard,
BoardDataStructure, and GameLogicPip. The BoardDataStructure represents
the backgammon game with a list of stacks. The GameLogicBoard adds
functionality on top of that with a) filling the board structure with
pips, b) moving a pip. The GameLogicPip class represents pip objects for
the board structure to use.

2. I find it's easier this way until it becomes clearer what packages
we'll break the program into.